+++ src/lib.rs
@@ -95,8 +95,9 @@ pub use crate::media::{
 pub use crate::pagination::PageRange;
 pub use crate::server::{PlexServer, ServerIdentity};
 pub use crate::traits::{
-    EditField, EditSummary, EditTags, EditTitle, FieldValue, HasCollections, HasGenres,
-    PlayedUnplayed, PlexObject, Ratable,
+    EditField, EditSummary, EditTags, EditTitle, FieldValue, HasArtLock, HasArtUrl, HasCollections,
+    HasGenres, HasPosterLock, HasPosterUrl, HasThemeLock, HasThemeUrl, PlayedUnplayed, PlexObject,
+    Ratable,
 };
 pub use crate::uri::PlexUri;
 pub use crate::util::{