+++ src/traits/mod.rs
@@ -17,11 +17,13 @@
 
 pub mod edit_field;
 pub mod edit_tags;
+pub mod images;
 pub mod played_unplayed;
 pub mod ratable;
 
 pub use edit_field::{EditField, EditSummary, EditTitle, FieldValue};
 pub use edit_tags::{EditTags, HasCollections, HasGenres};
+pub use images::{HasArtLock, HasArtUrl, HasPosterLock, HasPosterUrl, HasThemeLock, HasThemeUrl};
 pub use played_unplayed::PlayedUnplayed;
 pub use ratable::Ratable;